We made ItemEntry being a <int> value, so make all the usages of ItemEntry an Int value as well.

+++ The Noob Bot/Fisherbot/Main.cs
@@ -25,7 +25,7 @@ public class Main : IProduct
             if (string.IsNullOrWhiteSpace(FisherbotSetting.CurrentSetting.weaponName))
                 FisherbotSetting.CurrentSetting.weaponName =
                     ItemsManager.GetNameById(
-                        ObjectManager.Me.GetDescriptor<uint>(Descriptors.PlayerFields.VisibleItems + 15*2));
+                        (int) ObjectManager.Me.GetDescriptor<uint>(Descriptors.PlayerFields.VisibleItems + 15*2));
 
             Logging.Status = "Initialize Fisherbot Complete";
             Logging.Write("Initialize Fisherbot Complete");
